# Update label validation method to only run once per metric